AssetGuardianの主催者向け機能を呼び出す部分をOrganizerフォルダ内に移動した。
