+++ VitDeck/Assets/VitDeck/Organizer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b838f827962d8c04eb7d5a915ebc3f40
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
