+++ VitDeck/Assets/VitDeck/Organizer/AssetGuardianGUI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 50078d4d055c34645a3c4e0b683b9b19
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
