+++ VitDeck/Assets/VitDeck/Organizer/AssetGuardianGUI/AssetGuardianContextMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3dc81285b602f7f4da1d60c9e894c8cb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
