+++ VitDeck/Assets/VitDeck/Organizer/AssetGuardianGUI/VitDeck.Organizer.AssetGuardianGUI.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9018f4748dfe14f479da519849c0716c
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
